add hand
