+++ docker/shm-config.yaml
@@ -1,0 +1,3 @@
+services:
+  isaac-lab-base:
+    shm_size: '2gb'
