+++ tune_hand.bash
@@ -1,0 +1,2 @@
+ ./isaaclab.sh -p scripts/reinforcement_learning/ray/tuner.py   --cfg_file scripts/reinforcement_learning/ray/hyperparameter_tuning/hand_cfg.py   --cfg_class HandCfg   --run_mode local   --workflow scripts/reinforcement_learning/rl_games/train.py   --num_workers_per_node 4
+
